cleaning up, move transforms
